The last problem is solved. I have to just make the fields transient. The salebill is now generating but I cannot see the items in the sale Bill. An issues will be added for it.
A new UI for viewing all the salebill is to created. I added a DatePicker Jar and also changed some project settings so that the code can be started up faster, 

A new issue is to be created as we need a new SaleBill. If we press the save btn on the salebill again it creates a new salebill. Whereas it should modify the salebill